#33 refactor : NaverCloveClient를 리팩토링 한다
