+++ core/src/main/java/kusitms/duduk/core/newsletter/port/output/NaverClovaClientPort.java
@@ -1,0 +1,6 @@
+package kusitms.duduk.core.newsletter.port.output;
+
+public interface NaverClovaClientPort<T,R> {
+
+    R execute(T newsLetter);
+}
