--- core/src/main/java/kusitms/duduk/core/newsletter/port/output/NaverClovaSummaryClientPort.java
@@ -1,9 +1,0 @@
-package kusitms.duduk.core.newsletter.port.output;
-
-import kusitms.duduk.core.newsletter.dto.response.NaverClovaSummaryResponse;
-import kusitms.duduk.domain.newsletter.NewsLetter;
-
-public interface NaverClovaSummaryClientPort {
-
-    NaverClovaSummaryResponse summarize(NewsLetter newsLetter);
-}
