+++ docs/README.md
@@ -24,7 +24,7 @@
 - [x]  오늘의 뉴스레터에는 가장 최신의 에디터 뉴스레터를 보여준다
 - [x]  오늘의 단어는 오늘의 뉴스레터와 마찬가지로 에디터가 직접 작성한다
 - [x]  오늘의 뉴스레터와 오늘의 단어는 가장 최신의 날짜를 기준으로 조회한다
-- [ ]  실시간 트렌드 뉴스는 매 정각마다 크롤링한 뉴스로 대체한다 
+- [x]  실시간 트렌드 뉴스는 매 정각마다 크롤링한 뉴스로 대체한다 
 - [x]  실시간 트렌드 뉴스는 3개의 AI 뉴스 레터를 전달한다
 - [x]  추천 콘텐츠는 사용자의 관심사를 기반으로 추천한다 
 - [x]  추천 콘텐츠는 3개의 뉴스레터를 전달한다
@@ -32,7 +32,7 @@
 ### 뉴스레터
 **뉴스레터 생성**
 - [ ]  뉴스레터 생성 시 AI를 통해 투표와 설문이 생성된다
-  - [ ]  뉴스레터 생성 시 이벤트를 발행한다 
+  - [x]  뉴스레터 생성 시 이벤트를 발행한다 
   - [ ]  넘겨 받은 뉴스레터의 본문을 기반으로 투표와 설문을 생성한다 
   - [ ]  투표와 설문은 뉴스레터와 일대일 관계이다 
 
@@ -40,10 +40,6 @@
 - [x]  에디터는 썸네일, 헤드라인, 본문, 키워드, **카테고리(관심사)** 등을 입력하여 뉴스레터를 작성한다
 - [x]  에디터 권한이 있는 유저만 뉴스레터를 작성할 수 있다
 - [x]  에디터 권한은 ROLE_ADMIN, ROLE_EDITOR이다.
-- [ ]  에디터 본인이 작성한 뉴스레터에 대해 단어 하이퍼링크를 등록할 수 있다
-- [ ]  에디터 뉴스레터 헤드라인은 공백 포함 50자 미만이다
-- [ ]  에디터 뉴스레터 본문은 2000자를 넘을 수 없다
-- [ ]  뉴스레터 키워드는 최대 5개까지 등록 가능 하다
 
 **에디터 뉴스레터 조회**
 - [x]  오늘의 뉴스레터을 조회 한다
@@ -59,14 +55,14 @@
   - [x]  뉴스레터 상세 내용은 썸네일, 헤드라인, 키워드, 작성일, 본문, 코멘트 등의 정보를 확인할 수 있다
 - [x]  뉴스레터 상세 내용은 스크랩 여부를 확인할 수 있다
   - [x]  해당 정보는 비즈니스 로직 단에서 처리한다
-- [ ]  뉴스레터 상세 내용은 스크랩한 유저의 수를 확인할 수 있다
-  - [ ]  NewsLetter.getScrapCount().getValue()
-- [ ]  하나의 뉴스레터에 키워드는 최대 3개까지 보여진다
+- [x]  뉴스레터 상세 내용은 스크랩한 유저의 수를 확인할 수 있다
+  - [x]  NewsLetter.getScrapCount().getValue()
+- [x]  하나의 뉴스레터에 키워드는 최대 3개까지 보여진다
 - [x]  뉴스레터 클릭 시 조회수가 증가한다
   - [x]  RetrieveNewsLetterEvent가 발행된다
-- [ ]  사용자의 댓글 작성 여부를 확인 한 뒤 true/false 값을 반환한다 
-  - [ ]  findCommentByUserIdAndNewsLetterId() 메서드를 통해 사용자의 댓글을 확인한다
-  - [ ]  없다면 false 값을 반환한다 
+- [x]  사용자의 댓글 작성 여부를 확인 한 뒤 true/false 값을 반환한다 
+  - [x]  findCommentByUserIdAndNewsLetterId() 메서드를 통해 사용자의 댓글을 확인한다
+  - [x]  없다면 false 값을 반환한다 
 
 **에디터 뉴스레터 수정**
 - [ ]  에디터 본인이 작성한 뉴스레터만 수정이 가능하다
@@ -78,29 +74,29 @@
 - [ ]  삭제 시 해당 뉴스레터를 스크랩한 모든 유저에게서 해당 뉴스레터를 제거한다
 
 **AI 뉴스레터 생성**
-- [ ]  정해진 시간에 소스로 부터 뉴스를 크롤링 한다 
-- [ ]  카테고리 (관심사) 별 상위 5개의 뉴스를 크롤링 한다
-- [ ]  뉴스 내용을 API (네이버 클로바, Open AI)를 통해 뉴스레터 본문으로 정제한다
-    - [ ]  정제한 뉴스레터 본문을 데이터 베이스에 저장한다
-    - [ ]  AI 뉴스레터 본문은 2000자를 넘을 수 없다
-- [ ]  카테고리는 온보딩 시 사용자가 지정했던 관심사와 동일하다
-    - [ ]  사용자의 관심사와 동일한 뉴스는 별도의 메세지 큐로 전송한다
-- [ ] 썸네일 정보가 없다면 관심사 별 샘플 데이터를 랜덤으로 매핑한다
+- [x]  정해진 시간에 소스로 부터 뉴스를 크롤링 한다 
+- [x]  카테고리 (관심사) 별 상위 5개의 뉴스를 크롤링 한다
+- [x]  뉴스 내용을 API (네이버 클로바, Open AI)를 통해 뉴스레터 본문으로 정제한다
+    - [x]  정제한 뉴스레터 본문을 데이터 베이스에 저장한다
+    - [x]  AI 뉴스레터 본문은 2000자를 넘을 수 없다
+- [x]  카테고리는 온보딩 시 사용자가 지정했던 관심사와 동일하다
+    - [x]  사용자의 관심사와 동일한 뉴스는 별도의 메세지 큐로 전송한다
+- [x] 썸네일 정보가 없다면 크롤링 하지 않는다
 
 **뉴스레터 세줄 요약**
-- [ ]  뉴스레터 생성 시 CreateSummaryEvent를 발행한다
-- [ ]  해당 이벤트 생성 시 네이버 클로바 summary를 활용하여 뉴스레터를 세줄 요약하고 해당 정보를 데이터 베이스에 저장한다
-- [ ]  summary request 설정 값을 model = news, tone = 2, summaryCount = 3 으로 한다.
+- [x]  뉴스레터 생성 시 CreateSummaryEvent를 발행한다
+- [x]  해당 이벤트 생성 시 네이버 클로바 summary를 활용하여 뉴스레터를 세줄 요약하고 해당 정보를 데이터 베이스에 저장한다
+- [x]  summary request 설정 값을 model = news, tone = 2, summaryCount = 3 으로 한다.
 
 ### 코멘트 
 **뉴스레터 코멘트**
 - [x]  사용자는 뉴스레터에 코멘트를 남기면 다른 사람의 코멘트를 확인할 수 있다
 - [x]  코멘트의 글자 수는 300자 제한이다
 - [x]  코멘트에서 입력하는 정보는 전망 (긍정, 부정, 잘모르겠음) 및 내용이다
-- [ ]  다른 사람의 코멘트를 조회할 때는 좋아요 순으로 정렬한다 
-- [ ]  다른 사람의 코멘트에서 확인할 수 있는 정보는 닉네임, 전망, 내용, 좋아요 수, 작성일이다
-- [ ]  코멘트 작성 시 CreateCommentEvent를 발행한다
-- [ ]  CreateCommentEvent 발행 시 설문과 투표에 대한 접근 권한을 허용한다
+- [x]  다른 사람의 코멘트를 조회할 때는 좋아요 순으로 정렬한다 
+- [x]  다른 사람의 코멘트에서 확인할 수 있는 정보는 닉네임, 전망, 내용, 좋아요 수, 작성일이다
+- [x]  코멘트 작성 시 CreateCommentEvent를 발행한다
+- [x]  CreateCommentEvent 발행 시 설문과 투표에 대한 접근 권한을 허용한다
 
 ### 아카이브
 - [x]  사용자는 북마크를 통해 뉴스레터를 아카이브 할 수 있다
@@ -118,22 +114,15 @@
 **단어장 생성**
 - [ ]  관리자는 단어장을 추가할 수 있다
 - [x]  단어장은 단어, 영단어, 설명으로 구성되어 있다
-- [ ]  단어 설명은 300자를 넘어갈 수 없다
-
-**단어장 수정**
-- [ ]  관리자만 단어장을 수정할 수 있다
-
-**단어장 삭제**
-- [ ]  관리자만 단어장을 삭제할 수 있다
 
 ### 생각 더하기
 - [x]   코멘트를 달았던 뉴스레터들을 조회할 수 있다
   - [x]  뉴스 썸네일, 요약된 코멘트, 작성했던 코멘트 등을 반환한다
   - [x]  정렬 조건은 생각 구름 생성 여부, 작성일 순으로 한다 
-- [ ]  생각 구름을 만든다 
-  - [ ]  생각 구름은 사용자가 코멘트를 달았던 뉴스레터들을 기반으로 생성한다
-  - [ ]  요약하기 버튼을 누르면 요약을 진행한다 
-  - [ ]  다섯 가지의 생각을 하나의 생각 구름으로 묶는다
+- [x]  생각 구름을 만든다 
+  - [x]  생각 구름은 사용자가 코멘트를 달았던 뉴스레터들을 기반으로 생성한다
+  - [x]  요약하기 버튼을 누르면 요약을 진행한다 
+  - [x]  다섯 가지의 생각을 하나의 생각 구름으로 묶는다
 
 ### 생각 나누기 
 - [ ]  설문
@@ -143,8 +132,7 @@
 - [x]  사용자는 출결 현황을 확인할 수 있다
 - [x]  사용자는 카테고리 별 아카이브에서 북마크한 뉴스레터를 확인할 수 있다
 - [x]  사용자는 북마크한 단어장을 확인할 수 있다
-- [ ]  사용자는 닉네임을 변경할 수 있다 
-- [ ]  사용자는 프로필 사진을 변경할 수 있다 
+- [x]  사용자는 닉네임을 변경할 수 있다 
 
 ## 기술적 논의 사항
 Q. 성공적으로 로그인 했다면 이벤트를 발행하는건 어떨까?
